Modify part of asset classes
